Moved ion models
