+++ dids/src/main/kotlin/web5/sdk/dids/methods/ion/DidIon.kt
@@ -36,28 +36,28 @@ import web5.sdk.dids.DidMethod
 import web5.sdk.dids.DidResolutionMetadata
 import web5.sdk.dids.DidResolutionResult
 import web5.sdk.dids.ResolveDidOptions
-import web5.sdk.dids.ion.model.AddPublicKeysAction
-import web5.sdk.dids.ion.model.AddServicesAction
-import web5.sdk.dids.ion.model.Commitment
-import web5.sdk.dids.ion.model.DeactivateUpdateSignedData
-import web5.sdk.dids.ion.model.Delta
-import web5.sdk.dids.ion.model.Document
-import web5.sdk.dids.ion.model.InitialState
-import web5.sdk.dids.ion.model.OperationSuffixDataObject
-import web5.sdk.dids.ion.model.PatchAction
-import web5.sdk.dids.ion.model.PublicKey
-import web5.sdk.dids.ion.model.PublicKeyPurpose
-import web5.sdk.dids.ion.model.RecoveryUpdateSignedData
-import web5.sdk.dids.ion.model.RemovePublicKeysAction
-import web5.sdk.dids.ion.model.RemoveServicesAction
-import web5.sdk.dids.ion.model.ReplaceAction
-import web5.sdk.dids.ion.model.Reveal
-import web5.sdk.dids.ion.model.Service
-import web5.sdk.dids.ion.model.SidetreeCreateOperation
-import web5.sdk.dids.ion.model.SidetreeDeactivateOperation
-import web5.sdk.dids.ion.model.SidetreeRecoverOperation
-import web5.sdk.dids.ion.model.SidetreeUpdateOperation
-import web5.sdk.dids.ion.model.UpdateOperationSignedData
+import web5.sdk.dids.methods.ion.models.AddPublicKeysAction
+import web5.sdk.dids.methods.ion.models.AddServicesAction
+import web5.sdk.dids.methods.ion.models.Commitment
+import web5.sdk.dids.methods.ion.models.DeactivateUpdateSignedData
+import web5.sdk.dids.methods.ion.models.Delta
+import web5.sdk.dids.methods.ion.models.Document
+import web5.sdk.dids.methods.ion.models.InitialState
+import web5.sdk.dids.methods.ion.models.OperationSuffixDataObject
+import web5.sdk.dids.methods.ion.models.PatchAction
+import web5.sdk.dids.methods.ion.models.PublicKey
+import web5.sdk.dids.methods.ion.models.PublicKeyPurpose
+import web5.sdk.dids.methods.ion.models.RecoveryUpdateSignedData
+import web5.sdk.dids.methods.ion.models.RemovePublicKeysAction
+import web5.sdk.dids.methods.ion.models.RemoveServicesAction
+import web5.sdk.dids.methods.ion.models.ReplaceAction
+import web5.sdk.dids.methods.ion.models.Reveal
+import web5.sdk.dids.methods.ion.models.Service
+import web5.sdk.dids.methods.ion.models.SidetreeCreateOperation
+import web5.sdk.dids.methods.ion.models.SidetreeDeactivateOperation
+import web5.sdk.dids.methods.ion.models.SidetreeRecoverOperation
+import web5.sdk.dids.methods.ion.models.SidetreeUpdateOperation
+import web5.sdk.dids.methods.ion.models.UpdateOperationSignedData
 import java.net.URI
 import java.security.MessageDigest
 import java.util.UUID
